[refactor] 네비바 파일 내부 컴포넌트 분리
