+++ libs/components-shared/src/lib/Navbar.tsx
@@ -1,217 +1,24 @@
-import { CloseIcon, ExternalLinkIcon, HamburgerIcon, Icon } from '@chakra-ui/icons';
 import {
-  Avatar,
   Box,
-  Button,
-  ButtonProps,
   Container,
-  Divider,
-  Drawer,
-  DrawerBody,
-  DrawerCloseButton,
-  DrawerContent,
-  DrawerOverlay,
   Flex,
   Grid,
-  IconButton,
   Link,
-  Menu,
-  MenuButton,
-  MenuItem,
-  MenuList,
   Stack,
   useBreakpointValue,
-  useColorMode,
   useColorModeValue,
   useDisclosure,
 } from '@chakra-ui/react';
-import {
-  mainNavItems,
-  NavItem,
-  mypageNavLinks,
-  broadcasterCenterMypageNavLinks,
-} from '@project-lc/components-constants/navigation';
-import { ColorModeSwitcher } from '@project-lc/components-core/ColorModeSwitcher';
-import { useDisplaySize, useIsLoggedIn, useLogout, useProfile } from '@project-lc/hooks';
+import { mainNavItems } from '@project-lc/components-constants/navigation';
+import { useDisplaySize, useIsLoggedIn } from '@project-lc/hooks';
 import { UserType } from '@project-lc/shared-types';
 import NextLink from 'next/link';
 import { useRouter } from 'next/router';
-import React, { useCallback, useEffect, useMemo } from 'react';
-import { AiTwotoneSetting } from 'react-icons/ai';
-import { FaMoon, FaSun } from 'react-icons/fa';
+import React, { useCallback, useMemo } from 'react';
 import { KksLogo } from './KksLogo';
-import MypageNavbar from './MypageNavbar';
-import ProfileBox from './ProfileBox';
-import UserNotificationSection from './UserNotificationSection';
-
-/** 네비바 우측 버튼 공통 스타일 */
-export function NavbarRightButton(props: ButtonProps): JSX.Element {
-  const { children } = props;
-  return (
-    <Button
-      variant="unstyled"
-      minW={{ base: '50px', sm: '80px' }}
-      size="xs"
-      fontSize="1rem"
-      {...props}
-    >
-      {children}
-    </Button>
-  );
-}
-export function NavbarToggleButton({
-  onToggle,
-  isOpen,
-}: {
-  onToggle: () => void;
-  isOpen: boolean;
-}): JSX.Element {
-  return (
-    <IconButton
-      onClick={onToggle}
-      icon={isOpen ? <CloseIcon w={3} h={3} /> : <HamburgerIcon w={5} h={5} />}
-      variant="ghost"
-      aria-label="Toggle Navigation"
-    />
-  );
-}
-
-/** 네비바 우측영역(로그인 여부에 따라 로그인 버튼 혹은 프로필 사진이 표시됨) */
-export function NavbarRightButtonSection(): JSX.Element {
-  const router = useRouter();
-  const { isLoggedIn } = useIsLoggedIn();
-  return (
-    <Flex alignItems="center" justifyContent="flex-end">
-      {isLoggedIn ? (
-        // 로그인했을때
-        <>
-          <Box mr={{ base: '1', sm: '3' }}>
-            <UserNotificationSection />
-          </Box>
-
-          <PersonalPopoverMenu />
-        </>
-      ) : (
-        // 로그인 안했을때
-        <>
-          <ColorModeSwitcher />
-          {/* 로그인|회원가입 버튼그룹 컨테이너 */}
-          <Flex bg="blue.500" color="white" borderRadius="lg" height="32px" p={1}>
-            <NavbarRightButton onClick={() => router.push('/login')}>
-              로그인
-            </NavbarRightButton>
-            <Divider
-              display={{ base: 'none', md: 'inline-flex' }}
-              orientation="vertical"
-            />
-            <NavbarRightButton
-              onClick={() => router.push('/signup')}
-              display={{ base: 'none', md: 'inline-flex' }}
-            >
-              회원가입
-            </NavbarRightButton>
-          </Flex>
-        </>
-      )}
-    </Flex>
-  );
-}
-
-/** 로그인 한 경우에 사용하는 네비바 우측 개인메뉴 팝오버 */
-export function PersonalPopoverMenu(): JSX.Element {
-  const router = useRouter();
-  const { logout } = useLogout();
-  const { data: profileData } = useProfile();
-
-  const { colorMode, toggleColorMode } = useColorMode();
-  const SwitchIcon = useColorModeValue(FaMoon, FaSun);
-
-  const handleAccountSettingClick = useCallback(
-    () => router.push('/mypage/setting'),
-    [router],
-  );
-  return (
-    <Menu>
-      <MenuButton as={Avatar} size="sm" cursor="pointer" src={profileData?.avatar} />
-
-      <MenuList w={{ base: 280, sm: 300 }}>
-        {/* 프로필 표시 */}
-        <Box p={3}>
-          <ProfileBox allowAvatarChange />
-        </Box>
-        <Divider />
-
-        {/* 계정설정 버튼 */}
-        <MenuItem
-          my={1}
-          icon={<Icon fontSize="md" as={AiTwotoneSetting} />}
-          onClick={handleAccountSettingClick}
-        >
-          계정 설정
-        </MenuItem>
-
-        {/* 다크모드 버튼 */}
-        <MenuItem
-          my={1}
-          icon={<SwitchIcon />}
-          onClick={toggleColorMode}
-          closeOnSelect={false}
-        >
-          {colorMode === 'light' ? '다크모드' : '라이트모드'}
-        </MenuItem>
-
-        {/* 알림버튼 - 기존 알림버튼 누를시 팝오버로 알림을 표시했음. 별도 알림페이지 존재하지 않음
-        알림을 메뉴에 포함시켰을 때 알림목록 표시할 방법이 떠오르지 않아 개인메뉴에 포함시키지 않음
-        */}
-
-        {/* 로그아웃 버튼 */}
-        <MenuItem
-          my={1}
-          icon={<Icon fontSize="md" as={ExternalLinkIcon} />}
-          onClick={logout}
-        >
-          로그아웃
-        </MenuItem>
-      </MenuList>
-    </Menu>
-  );
-}
-
-/** 네비바에 사용하는 크크쇼 로고(링크) */
-export function NavbarLinkLogo({ appType }: { appType: UserType }): JSX.Element {
-  return (
-    <NextLink href="/" passHref>
-      <Link
-        textAlign={useBreakpointValue({ base: 'center', md: 'left' })}
-        color={useColorModeValue('gray.800', 'white')}
-      >
-        <KksLogo appType={appType} size="small" />
-      </Link>
-    </NextLink>
-  );
-}
-
-/**  모바일 화면에서 네비바 레이아웃 ******* 로고 가운데 위치시키기 위해 Grid 사용 */
-function MobileNavLayout({ children }: { children: React.ReactNode }): JSX.Element {
-  return (
-    <Grid display={{ base: 'grid', md: 'none' }} templateColumns="repeat(3, 1fr)">
-      {children}
-    </Grid>
-  );
-}
-
-/** 데스크톱 화면에서 네비바 레이아웃 ****** Flex로 표시 */
-function DesktopNavLayout({ children }: { children: React.ReactNode }): JSX.Element {
-  return (
-    <Flex
-      justifyContent="space-between"
-      alignItems="center"
-      display={{ base: 'none', md: 'flex' }}
-    >
-      {children}
-    </Flex>
-  );
-}
+import { MobileSideDrawerNav } from './navbar/MobileSideDrawerNav';
+import { NavbarRightButtonSection } from './navbar/NavbarRightButtonSection';
+import { NavbarToggleButton } from './navbar/NavbarToggleButton';
 
 export interface NavbarProps {
   appType?: UserType;
@@ -225,19 +32,13 @@ export function Navbar({ appType = 'seller' }: NavbarProps): JSX.Element {
   const hambergerButton = useMemo(() => {
     return (
       <Flex alignItems="center">
-        <IconButton
-          ml={-2}
-          onClick={onToggle}
-          icon={isOpen ? <CloseIcon w={3} h={3} /> : <HamburgerIcon w={5} h={5} />}
-          variant="ghost"
-          aria-label="Toggle Navigation"
-        />
+        <NavbarToggleButton onToggle={onToggle} isOpen={isOpen} />
       </Flex>
     );
   }, [isOpen, onToggle]);
 
   return (
-    // {/* 배경이 되는 wrapper 컴포넌트 */}
+    // {/* Navbar 배경이 되는 wrapper 컴포넌트 */}
     <Box
       bg={useColorModeValue('white', 'gray.800')}
       color={useColorModeValue('gray.600', 'white')}
@@ -270,8 +71,36 @@ export function Navbar({ appType = 'seller' }: NavbarProps): JSX.Element {
       </Container>
 
       {/* 토글버튼(hambergerButton)으로 여닫는 모바일화면 네비바(Drawer) */}
-      <MobileDrawerNav isOpen={isOpen} onClose={onClose} appType={appType} />
+      <MobileSideDrawerNav isOpen={isOpen} onClose={onClose} appType={appType} />
     </Box>
+  );
+}
+
+/** 네비바에 사용하는 크크쇼 로고(링크) */
+export function NavbarLinkLogo({ appType }: { appType: UserType }): JSX.Element {
+  return (
+    <NextLink href="/" passHref>
+      <Link
+        textAlign={useBreakpointValue({ base: 'center', md: 'left' })}
+        color={useColorModeValue('gray.800', 'white')}
+      >
+        <KksLogo appType={appType} size="small" />
+      </Link>
+    </NextLink>
+  );
+}
+
+/**  모바일 화면에서 네비바 레이아웃 ******* 로고 가운데 위치시키기 위해 Grid 사용 */
+function MobileNavLayout({ children }: { children: React.ReactNode }): JSX.Element {
+  return <Grid templateColumns="repeat(3, 1fr)">{children}</Grid>;
+}
+
+/** 데스크톱 화면에서 네비바 레이아웃 ****** Flex로 표시 */
+function DesktopNavLayout({ children }: { children: React.ReactNode }): JSX.Element {
+  return (
+    <Flex justifyContent="space-between" alignItems="center">
+      {children}
+    </Flex>
   );
 }
 
@@ -319,102 +148,6 @@ const MainNavItemList = (): JSX.Element => {
         </Box>
       ))}
     </Stack>
-  );
-};
-
-/** 모바일 화면 네비게이션(좌측 Drawer) */
-const MobileDrawerNav = ({
-  isOpen,
-  onClose,
-  appType,
-}: {
-  isOpen: boolean;
-  onClose: () => void;
-  appType: UserType;
-}): JSX.Element => {
-  const { isLoggedIn } = useIsLoggedIn();
-  const router = useRouter();
-  const { isMobileSize } = useDisplaySize();
-
-  useEffect(() => {
-    if (!isMobileSize) onClose();
-  }, [isMobileSize, onClose]);
-
-  return (
-    <Drawer isOpen={isOpen} placement="left" size="xs" onClose={onClose}>
-      <DrawerOverlay />
-      <DrawerContent bgGradient="linear(to-r, blue.300, blue.500)" color="white" pt={8}>
-        <DrawerCloseButton />
-
-        <DrawerBody>
-          <Stack>
-            {mainNavItems.map((navItem) => (
-              <MobileNavItem key={navItem.label} {...navItem} onClose={onClose} />
-            ))}
-            {/* 로그인 한 경우 마이페이지 nav 메뉴를 표시함 */}
-            {isLoggedIn && (
-              <MypageNavbar
-                navLinks={
-                  appType === 'seller' ? mypageNavLinks : broadcasterCenterMypageNavLinks
-                }
-              />
-            )}
-
-            {/* 로그인 하지 않은 경우에만 드로어에 회원가입 버튼을 표시한다 */}
-            {!isLoggedIn && (
-              <Button
-                onClick={() => {
-                  router.push('/signup');
-                  onClose();
-                }}
-                textAlign="left"
-                fontSize="xl"
-                fontWeight="bold"
-                fontFamily="Gmarket Sans"
-                _hover={{ textDecoration: 'underline' }}
-                variant="unstyled"
-              >
-                회원가입
-              </Button>
-            )}
-          </Stack>
-        </DrawerBody>
-      </DrawerContent>
-    </Drawer>
-  );
-};
-
-const MobileNavItem = ({
-  label,
-  href,
-  isExternal,
-  needLogin,
-  onClose,
-}: NavItem & { onClose: () => void }): JSX.Element | null => {
-  const { isLoggedIn } = useIsLoggedIn();
-  const router = useRouter();
-
-  if (needLogin && !isLoggedIn) return null;
-
-  return (
-    <Button
-      textAlign="left"
-      fontSize="xl"
-      fontWeight="bold"
-      fontFamily="Gmarket Sans"
-      _hover={{ textDecoration: 'underline' }}
-      variant="unstyled"
-      onClick={() => {
-        if (isExternal) {
-          window.open(href, '_blank');
-        } else {
-          router.push(href);
-        }
-        onClose();
-      }}
-    >
-      {label}
-    </Button>
   );
 };
